Changes made to email notifications
By default BB doesn't email a user if they are assigned a ticket. It
also doesn't email a person if a ticket they are assigned to is
commented on. I also wanted everyone who is apart of the group
"plexmins" to be emailed when any "new" issue was logged.

There is an issue where if you comment on a ticket you are assigned to
you will still get an email.